Extract pure utility helpers into src/core/normalize.ts
Move username/subreddit sanitizers, version parsing, timestamp and
generic normalize helpers into a leaf module. Typecheck, 167 tests,
and build all pass.
